Refactored feature tests into helper method
