+++ spec/spec_helper.rb
@@ -32,7 +32,7 @@ RSpec.configure do |config|
     # ...rather than:
     #     # => "be bigger than 2"
     expectations.include_chain_clauses_in_custom_matcher_descriptions = true
-    config.include Helpers::Authentication, type: :feature
+    #config.include Helpers::Authentication, type: :feature
   end
 
   # rspec-mocks config goes here. You can use an alternate test double
